refactor(actor-sidebar): rename file as it doesn't do anything with sidebar
